+++ pysat/_instrument.py
@@ -7,17 +7,20 @@ try:
 except NameError:
     basestring = str
 
-import string
-import os
 import copy
-import sys
 import datetime as dt
+import functools
+import inspect
+import os
+import string
+import sys
 
 import pkgutil
-import pandas as pds
-import numpy as np
-import xarray as xr
 import warnings
+
+import numpy as np
+import pandas as pds
+import xarray as xr
 
 from pysat import _custom
 from pysat import _files
@@ -356,6 +359,9 @@ class Instrument(object):
         self._export_meta_post_processing = None
 
         # store kwargs, passed to load routine
+        # first, check if keywords are  valid
+        _check_if_keywords_supported(self._load_rtn, **kwargs)
+        # store
         self.kwargs = kwargs
 
         # run instrument init function, a basic pass function is used
@@ -371,40 +377,6 @@ class Instrument(object):
                           ' instruments. (strict_time_flag)',
                           DeprecationWarning, stacklevel=2)
 
-    def __setattr__(self, name, value):
-        """Moves instrument attributes onto meta attributes
-
-        If the attribute is not in _base_attrs, add to meta attributes.
-        For all other cases, store as an instrument attribute.
-        """
-
-        if '_base_attr' in dir(self):
-            if name not in self._base_attr:
-                # set attribute on meta
-                if name[0] != '_':
-                    object.__setattr__(self.meta, name, value)
-                else:
-                    object.__setattr__(self, name, value)
-            else:
-                object.__setattr__(self, name, value)
-        else:
-            object.__setattr__(self, name, value)
-
-    def __getattr__(self, name):
-        """Gets instrument attributes from meta attributes
-
-        Usually, python only calls __getattr__ if name does not already
-        exist in the instrument, so we only need to check
-        the meta object. However, __copy__ calls __getattr__, so we still have
-        to check for invalid attributes manually.
-        """
-        if name not in self.__dict__:
-            try:
-                return getattr(self.meta, name)
-            except:
-                raise AttributeError("No attribute {}".format(name))
-
-        return getattr(self.meta, name)
 
     def __getitem__(self, key):
         """
@@ -1442,6 +1414,7 @@ class Instrument(object):
 
         # transfer any extra attributes in meta to the Instrument object
         self.meta.transfer_attributes_to_instrument(self)
+        self.meta.mutable = False
         sys.stdout.flush()
         return
 
@@ -2606,3 +2579,105 @@ class Instrument(object):
             # attach attributes
             out_data.setncatts(adict)
         return
+
+#
+# ----------------------------------------------
+#   Utilities supporting the Instrument Object
+# ----------------------------------------------
+#
+
+
+def _get_supported_keywords(load_func):
+    """Return a list of supported keywords
+
+    Intended to be used on the supporting instrument
+    functions that enable the general Instrument object
+    to load and work with a particular data set.
+
+    Parameters
+    ----------
+    load_func: Python method or functools.partial
+        Method used to load data within pysat
+
+    Returns
+    -------
+    list
+        list of keyword argument strings
+
+
+    Notes
+    -----
+        If the input is a partial function then the
+        list of keywords returned only includes keywords
+        that have not already been set as part of
+        the functools.partial instantiation.
+
+    """
+
+    # check if partial function
+    if isinstance(load_func, functools.partial):
+        # get keyword arguments already applied to function
+        existing_kws = load_func.keywords
+        # pull out python function portion
+        load_func = load_func.func
+    else:
+        existing_kws = None
+
+    # modified from code on
+    # https://stackoverflow.com/questions/196960/can-you-list-the-keyword-arguments-a-function-receives
+    if sys.version_info.major == 2:
+        args, varargs, varkw, defaults = inspect.getargspec(load_func)
+    else:
+        sig = inspect.getfullargspec(load_func)
+        # args are first
+        args = sig.args
+
+    pop_list = []
+    # account for keywords that exist for every load function
+    pre_kws = ['fnames', 'sat_id', 'tag']
+    # account for keywords already set since input was a partial function
+    if existing_kws is not None:
+        pre_kws.extend(existing_kws.keys())
+    # remove pre-existing keywords from output
+    # first identify locations
+    for i, arg in enumerate(args):
+        if arg in pre_kws:
+            pop_list.append(i)
+    # remove identified locations
+    # go backwards so we don't mess with the location of data we
+    # are trying to remove
+    if len(pop_list) > 0:
+        for pop in pop_list[::-1]:
+            args.pop(pop)
+
+    # *args and **kwargs are not required, so ignore them.
+    return args
+
+
+def _check_if_keywords_supported(func, **kwargs):
+    """Checks if keywords supported by function
+
+    Parameters
+    ----------
+    func: method
+        Method to be checked against
+    **kwargs : keyword args
+        keyword arguments dictionary
+
+    Returns
+    -------
+    bool
+        If true, all keywords supported
+
+    """
+
+    # get list of supported keywords
+    supp = _get_supported_keywords(func)
+    # check if kwargs are in list
+    for name in kwargs.keys():
+        if name not in supp:
+            estr = ' '.join((name, 'is not a supported keyword by pysat or',
+                             'by the underlying supporting load routine.',
+                             'Please double check the keyword inputs.'))
+            raise ValueError(estr)
+    return True