+++ pysat/instruments/pysat_testing_xarray.py
@@ -4,15 +4,14 @@ Produces fake instrument data for testing.
 """
 from __future__ import print_function
 from __future__ import absolute_import
-import os
-
 import datetime as dt
+import functools
 import numpy as np
-import pandas as pds
+
 import xarray
 
 import pysat
-from pysat.instruments.methods import testing as test
+from pysat.instruments.methods import testing as mm_test
 
 # pysat required parameters
 platform = 'pysat'
@@ -26,7 +25,46 @@ pandas_format = False
 
 
 def init(self):
+    """Initializes the Instrument object with instrument specific values.
+
+    Runs once upon instantiation.
+
+    Parameters
+    ----------
+    self : pysat.Instrument
+        This object
+
+    Returns
+    --------
+    Void : (NoneType)
+        Object modified in place.
+
+
+    """
+
     self.new_thing = True
+
+
+def default(inst):
+    """Default customization function.
+
+    This routine is automatically applied to the Instrument object
+    on every load by the pysat nanokernel (first in queue).
+
+    Parameters
+    ----------
+    self : pysat.Instrument
+        This object
+
+    Returns
+    --------
+    Void : (NoneType)
+        Object modified in place.
+
+
+    """
+
+    pass
 
 
 def load(fnames, tag=None, sat_id=None, sim_multi_file_right=False,
@@ -42,7 +80,7 @@ def load(fnames, tag=None, sat_id=None, sim_multi_file_right=False,
         Instrument tag (accepts '')
     sat_id : (str or NoneType)
         Instrument satellite ID (accepts '' or a number (i.e., '10'), which
-        specifies the number of times to include in the test instrument)
+        specifies the number of data points to include in the test instrument)
     sim_multi_file_right : (boolean)
         Adjusts date range to be 12 hours in the future or twelve hours beyond
         root_date (default=False)
@@ -60,33 +98,24 @@ def load(fnames, tag=None, sat_id=None, sim_multi_file_right=False,
     data : (xr.Dataset)
         Testing data
     meta : (pysat.Meta)
-        Metadataxs
+        Metadata
 
     """
 
     # create an artifical satellite data set
-    parts = os.path.split(fnames[0])[-1].split('-')
-    yr = int(parts[0])
-    month = int(parts[1])
-    day = int(parts[2][0:2])
+    iperiod = mm_test.define_period()
+    drange = mm_test.define_range()
+    uts, index, date = mm_test.generate_times(fnames, sat_id=sat_id, freq='1S')
 
-    date = dt.datetime(yr, month, day)
     if sim_multi_file_right:
         root_date = dt.datetime(2009, 1, 1, 12)
-        data_date = date + pds.DateOffset(hours=12)
     elif sim_multi_file_left:
         root_date = dt.datetime(2008, 12, 31, 12)
-        data_date = date - pds.DateOffset(hours=12)
     else:
         root_date = dt.datetime(2009, 1, 1)
-        data_date = date
-    num = 86400 if sat_id == '' else int(sat_id)
-    num_array = np.arange(num)
-    index = pds.date_range(data_date,
-                           data_date+pds.DateOffset(seconds=num-1),
-                           freq='S')
+
     if malformed_index:
-        index = index[0:num].tolist()
+        index = index.tolist()
         # nonmonotonic
         index[0:3], index[3:6] = index[3:6], index[0:3]
         # non unique
@@ -96,34 +125,37 @@ def load(fnames, tag=None, sat_id=None, sim_multi_file_right=False,
     # need to create simple orbits here. Have start of first orbit
     # at 2009,1, 0 UT. 14.84 orbits per day
     time_delta = date - root_date
-    mlt = test.generate_fake_data(time_delta.total_seconds(), num_array,
-                                  period=5820, data_range=[0.0, 24.0])
+    mlt = mm_test.generate_fake_data(time_delta.total_seconds(), uts,
+                                     period=iperiod['lt'],
+                                     data_range=drange['lt'])
     data['mlt'] = (('time'), mlt)
 
     # do slt, 20 second offset from mlt
-    slt = test.generate_fake_data(time_delta.total_seconds()+20, num_array,
-                                  period=5820, data_range=[0.0, 24.0])
+    slt = mm_test.generate_fake_data(time_delta.total_seconds()+20, uts,
+                                     period=iperiod['lt'],
+                                     data_range=drange['lt'])
     data['slt'] = (('time'), slt)
 
     # create a fake longitude, resets every 6240 seconds
     # sat moves at 360/5820 deg/s, Earth rotates at 360/86400, takes extra time
     # to go around full longitude
-    longitude = test.generate_fake_data(time_delta.total_seconds(), num_array,
-                                        period=6240, data_range=[0.0, 360.0])
+    longitude = mm_test.generate_fake_data(time_delta.total_seconds(), uts,
+                                           period=iperiod['lon'],
+                                           data_range=drange['lon'])
     data['longitude'] = (('time'), longitude)
 
     # create latitude area for testing polar orbits
-    angle = test.generate_fake_data(time_delta.total_seconds(),
-                                    num_array, period=5820,
-                                    data_range=[0.0, 2.0*np.pi])
+    angle = mm_test.generate_fake_data(time_delta.total_seconds(), uts,
+                                       period=iperiod['angle'],
+                                       data_range=drange['angle'])
     latitude = 90.0 * np.cos(angle)
     data['latitude'] = (('time'), latitude)
 
     # fake orbit number
     fake_delta = date - dt.datetime(2008, 1, 1)
-    orbit_num = test.generate_fake_data(fake_delta.total_seconds(),
-                                        num_array, period=5820,
-                                        cyclic=False)
+    orbit_num = mm_test.generate_fake_data(fake_delta.total_seconds(),
+                                           uts, period=iperiod['lt'],
+                                           cyclic=False)
 
     data['orbit_num'] = (('time'), orbit_num)
 
@@ -133,7 +165,7 @@ def load(fnames, tag=None, sat_id=None, sim_multi_file_right=False,
     data['dummy1'] = (('time'), mlt_int)
     data['dummy2'] = (('time'), long_int)
     data['dummy3'] = (('time'), mlt_int + long_int * 1000.)
-    data['dummy4'] = (('time'), num_array)
+    data['dummy4'] = (('time'), uts)
     data['string_dummy'] = (('time'), ['test'] * len(data.indexes['time']))
     data['unicode_dummy'] = (('time'), [u'test'] * len(data.indexes['time']))
     data['int8_dummy'] = (('time'), np.array([1] * len(data.indexes['time']),
@@ -148,18 +180,8 @@ def load(fnames, tag=None, sat_id=None, sim_multi_file_right=False,
     return data, meta.copy()
 
 
-def list_files(tag=None, sat_id=None, data_path=None, format_str=None):
-    """Produce a fake list of files spanning a year"""
-
-    index = pds.date_range(dt.datetime(2008, 1, 1),
-                           dt.datetime(2010, 12, 31))
-    names = [data_path+date.strftime('%Y-%m-%d')+'.nofile' for date in index]
-    return pds.Series(names, index=index)
-
-
-def download(date_array, tag, sat_id, data_path=None, user=None,
-             password=None):
-    pass
+list_files = functools.partial(mm_test.list_files, test_dates=_test_dates)
+download = functools.partial(mm_test.download)
 
 
 meta = pysat.Meta()